Remove console log and move useInstallPWA into it's own file
